Clean up include order fix
